backend/requirements.txt
# Core API Framework
fastapi
uvicorn[standard]

# Data Validation & Structured Output
pydantic>=2.0.0

# Environment Variables
python-dotenv

# AI Engine
google-genai

# Production Server (optional, for gunicorn deployment)
gunicorn